1. still return filepath even if xxx in ImageUtil.compressImage 2. return Long instead of Integer in RandomUtil.digitsStartWithZero 3. fix show tables and show creation for sqlserver

+++ sirap-basic/src/main/java/com/sirap/basic/util/RandomUtil.java
@@ -10,11 +10,11 @@ public class RandomUtil {
 		return result;
 	}
 
-	public static Integer digitsStartWithNoZero(int countOfChars) {
+	public static Long digitsStartWithNoZero(int countOfChars) {
 		String start = chars(1, "123456798");
 		String others = chars(countOfChars - 1, StrUtil.DIGITS);
 		
-		return Integer.parseInt(start + others);
+		return Long.parseLong(start + others);
 	}
 
 	public static String letters(int countOfChars) {
